MinRelSize and RelWithDebInfo supported properly by run.sh

+++ run.sh
@@ -1,48 +1,52 @@
 #!/bin/bash
 
-build_type=Release  # Conan default
-valgrind=memcheck   # CTest default
+conan_config=Release
+cmake_config=Release
+valgrind=memcheck
 
 while (( $# )); do
     case $1 in
     Clean)
         clean=1
-        offset=1
+        shift 1
         ;;
-    Debug|Release|RelWithDebInfo|MinSizeRel)
-        build_type=$1
-        offset=1
+    Debug)
+        conan_config=Debug
+        cmake_config=$1
+        shift 1
+        ;;
+    Release|MinSizeRel|RelWithDebInfo)
+        conan_config=Release
+        cmake_config=$1
+        shift 1
         ;;
     Coverage)
         coverage=$2
-        offset=2
+        shift 2
         ;;
     MemCheck)
         memcheck=1
         valgrind=$2
-        offset=2
+        shift 2
         ;;
-    Sanitize)
-        sanitize=$2
-        offset=2
+    Sanitizer)
+        sanitizer=$2
+        shift 2
         ;;
     *)
         echo "unknown option '$1'"
         exit 1
     esac
-    shift $offset
 done
 
-build_dir=build/$build_type
+build_dir=build/$cmake_config
 make_cmd="make -C $build_dir --no-print-directory"
 
 set -e
-if [[ ! -z $clean ]]; then
-    rm -rf build
-fi
+[[ -z $clean ]] || rm -rf build
 mkdir -p $build_dir
-conan install -s build_type=$build_type -s compiler.libcxx=libstdc++11 -if $build_dir .
-cmake -DCMAKE_BUILD_TYPE=$build_type -Dprojname_coverage=$coverage -Dprojname_valgrind=$valgrind -Dprojname_sanitizer=$sanitize -B$build_dir -H.
+conan install -s build_type=$conan_config -s compiler.libcxx=libstdc++11 -if $build_dir .
+cmake -DCMAKE_BUILD_TYPE=$cmake_config -Dprojname_coverage=$coverage -Dprojname_valgrind=$valgrind -Dprojname_sanitizer=$sanitizer -B$build_dir -H.
 $make_cmd format
 $make_cmd all
 if [[ ! -z $memcheck ]]; then
